Move functions to seperate file
